Make POSBookingDateFormatter
