+++ Modules/Sources/PointOfSale/Presentation/Bookings/CancelBooking/POSCancelBookingModalContent.swift
@@ -7,7 +7,6 @@ struct POSCancelBookingModalContent: View {
     @Binding var cancelModalState: CancelBookingModalState?
 
     @Environment(POSBookingsModel.self) private var bookingsModel
-    @Environment(\.siteTimezone) private var siteTimezone
 
     var body: some View {
         switch state {
@@ -15,7 +14,7 @@ struct POSCancelBookingModalContent: View {
             POSCancelBookingConfirmationView(
                 bookingNumber: booking.id,
                 serviceName: booking.serviceName,
-                formattedDateTime: formattedCancelDateTime,
+                formattedDateTime: POSBookingDateFormatter.formattedDateTime(for: booking.startDate),
                 customerName: booking.customerName,
                 isProcessing: false,
                 onClose: { cancelModalState = nil },
@@ -31,7 +30,7 @@ struct POSCancelBookingModalContent: View {
             POSCancelBookingConfirmationView(
                 bookingNumber: booking.id,
                 serviceName: booking.serviceName,
-                formattedDateTime: formattedCancelDateTime,
+                formattedDateTime: POSBookingDateFormatter.formattedDateTime(for: booking.startDate),
                 customerName: booking.customerName,
                 isProcessing: true,
                 onClose: {},
@@ -57,19 +56,6 @@ struct POSCancelBookingModalContent: View {
                 onClose: { cancelModalState = nil }
             )
         }
-    }
-
-    private static let cancelDateTimeFormatter: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.dateStyle = .medium
-        formatter.timeStyle = .short
-        return formatter
-    }()
-
-    private var formattedCancelDateTime: String {
-        // Use UTC: the API timestamps already represent site-local wall-clock time.
-        Self.cancelDateTimeFormatter.timeZone = TimeZone(identifier: "UTC")
-        return Self.cancelDateTimeFormatter.string(from: booking.startDate)
     }
 
     @MainActor
