+++ Modules/Sources/PointOfSale/Presentation/Bookings/POSBookingRowView.swift
@@ -7,7 +7,6 @@ struct POSBookingRowView: View {
 
     @ScaledMetric private var scale: CGFloat = 1.0
     @Environment(\.dynamicTypeSize) var dynamicTypeSize
-    @Environment(\.siteTimezone) private var siteTimezone
 
     var body: some View {
         VStack(alignment: .leading, spacing: POSSpacing.none) {
@@ -34,7 +33,7 @@ struct POSBookingRowView: View {
     @ViewBuilder
     private var bookingHeaderRow: some View {
         HStack(alignment: .center) {
-            Text(POSBookingSummaryView.formattedTimeRange(for: booking, siteTimezone: siteTimezone))
+            Text(POSBookingDateFormatter.formattedTimeRange(for: booking))
                 .font(.posBodySmallBold())
                 .foregroundStyle(Color.posOnSurface)
                 .fixedSize(horizontal: false, vertical: true)
@@ -46,12 +45,9 @@ struct POSBookingRowView: View {
     }
 
     private var accessibilityLabel: String {
-        let formatter = DateFormatter.posAccessibilityTimeFormatter
-        // Use UTC: the API timestamps already represent site-local wall-clock time.
-        formatter.timeZone = TimeZone(identifier: "UTC")
         let timeRange = Localization.timeRangeAccessibilityLabel(
-            start: formatter.string(from: booking.startDate),
-            end: formatter.string(from: booking.endDate)
+            start: POSBookingDateFormatter.accessibilityFormattedTime(for: booking.startDate),
+            end: POSBookingDateFormatter.accessibilityFormattedTime(for: booking.endDate)
         )
 
         let customerDisplayName = booking.customerName ?? booking.customerEmail
@@ -78,13 +74,4 @@ private enum Localization {
         )
         return String(format: format, start, end)
     }
-}
-
-private extension DateFormatter {
-    static let posAccessibilityTimeFormatter: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.dateStyle = .none
-        formatter.timeStyle = .short
-        return formatter
-    }()
 }
